+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpVerifier.kt
@@ -56,6 +56,9 @@ open class OpenId4VpVerifier(
     private val responseParser = ResponseParser(jwsService, verifierJwsService)
     private val timeLeeway = timeLeewaySeconds.toDuration(DurationUnit.SECONDS)
     private val supportedAlgorithms = verifierJwsService.supportedAlgorithms.map { it.identifier }
+    private val supportedSignatureVerificationAlgorithm =
+        (verifierJwsService.supportedAlgorithms.firstOrNull { it == JwsAlgorithm.ES256 }?.identifier
+            ?: verifierJwsService.supportedAlgorithms.first().identifier)
     private val containerJwt = FormatContainerJwt(algorithmStrings = supportedAlgorithms)
     private val containerSdJwt = FormatContainerSdJwt(
         sdJwtAlgorithmStrings = supportedAlgorithms.toSet(),
@@ -81,6 +84,7 @@ open class OpenId4VpVerifier(
         RelyingPartyMetadata(
             redirectUris = listOfNotNull((clientIdScheme as? ClientIdScheme.RedirectUri)?.redirectUri),
             jsonWebKeySet = JsonWebKeySet(listOf(keyMaterial.publicKey.toJsonWebKey())),
+            authorizationSignedResponseAlgString = supportedSignatureVerificationAlgorithm,
             subjectSyntaxTypesSupported = setOf(
                 OpenIdConstants.URN_TYPE_JWK_THUMBPRINT,
                 OpenIdConstants.PREFIX_DID_KEY,
@@ -281,7 +285,7 @@ open class OpenId4VpVerifier(
         clientId = clientIdScheme.clientId,
         redirectUrl = if (!requestOptions.isAnyDirectPost) clientIdScheme.redirectUri else null,
         responseUrl = requestOptions.responseUrl,
-        scope = requestOptions.buildScope(),
+        //scope = requestOptions.buildScope(), // TODO verify if this is needed
         nonce = nonceService.provideNonce(),
         walletNonce = requestObjectParameters?.walletNonce,
         clientMetadata = clientMetadata(requestOptions),
@@ -326,15 +330,13 @@ open class OpenId4VpVerifier(
     /**
      * Validates an Authentication Response from the Wallet, where [input] is a map of POST parameters received.
      */
-    suspend fun validateAuthnResponse(input: Map<String, String>): AuthnResponseResult {
-        val paramsFrom = runCatching {
+    suspend fun validateAuthnResponse(input: Map<String, String>): AuthnResponseResult =
+        runCatching {
             ResponseParametersFrom.Post(input.decode<AuthenticationResponseParameters>())
         }.getOrElse {
             Napier.w("Could not parse authentication response: $input", it)
             return AuthnResponseResult.Error("Can't parse input", null)
-        }
-        return validateAuthnResponse(paramsFrom)
-    }
+        }.let { validateAuthnResponse(it) }
 
     /**
      * Validates an Authentication Response from the Wallet, where [input] is either:
@@ -342,15 +344,15 @@ open class OpenId4VpVerifier(
      * - a URL, containing parameters in the query, e.g. `https://example.com?id_token=...`
      * - parameters encoded as a POST body, e.g. `id_token=...&vp_token=...`
      */
-    suspend fun validateAuthnResponse(input: String): AuthnResponseResult {
-        val paramsFrom = runCatching {
+    suspend fun validateAuthnResponse(input: String): AuthnResponseResult =
+        runCatching {
             responseParser.parseAuthnResponse(input)
         }.getOrElse {
             Napier.w("Could not parse authentication response: $input", it)
             return AuthnResponseResult.Error("Can't parse input", null)
+        }.let {
+            validateAuthnResponse(it)
         }
-        return validateAuthnResponse(paramsFrom)
-    }
 
     /**
      * Validates [AuthenticationResponseParameters] from the Wallet