+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/AuthnResponseResult.kt
@@ -12,12 +12,12 @@ sealed class AuthnResponseResult {
     /**
      * Error in parsing the URL or content itself, before verifying the contents of the OpenId response
      */
-    data class Error(val reason: String, val state: String?) : AuthnResponseResult()
+    data class Error(val reason: String, val state: String?, val cause: Throwable? = null) : AuthnResponseResult()
 
     /**
      * Error when validating the `vpToken` or `idToken`
      */
-    data class ValidationError(val field: String, val state: String?) : AuthnResponseResult()
+    data class ValidationError(val field: String, val state: String?, val cause: Throwable? = null) : AuthnResponseResult()
 
     /**
      * Wallet provided an `id_token`, no `vp_token` (as requested by us!)